requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow
numpy
opencv-python-headless
